requirements.txt
anthropic>=0.40.0
rich>=13.0.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.32.0
jinja2>=3.1.0
python-multipart>=0.0.18
playwright>=1.54.0
twilio>=9.3.0
